restruct 0.7 based 1.6
弃用upload.py
upload_userinfo和upload_address放入project/roles/user.py

upload_address改进代码building = Building.objects.get(id='0000000000')
